modificaciones numeroticket

+++ src/entidades/caja/VentaCobranza.java
@@ -25,7 +25,7 @@ public class VentaCobranza implements Comparator {
     private Date fecha;
     private BigDecimal importe;
     private BigDecimal importeSaldo;
-    private Long numero;
+    private int numero;
     private BigDecimal saldoAcumulado;
     private String observaciones;
 //Franquito: get y set objeto
@@ -72,11 +72,11 @@ public class VentaCobranza implements Comparator {
         this.importeSaldo = importeSaldo;
     }
 
-    public Long getNumero() {
+    public int getNumero() {
         return numero;
     }
 
-    public void setNumero(Long numero) {
+    public void setNumero(int numero) {
         this.numero = numero;
     }
